+++ src/index.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './components/App';
+
+
 
 ReactDOM.render(
   <React.StrictMode>
